+++ test/form.test.js
@@ -1178,6 +1178,7 @@ describe('Form:', sandbox(function () {
                         id: 'foo',
                         template: 'form.test.jade',
                         setValues: this.stub().callsArg(2),
+                        preRedirect: this.stub().callsArg(2),
                         getValues: function (req, res, callback) {
                             callback(null, { firstName: 'FooFoo', lastName: 'BarBar', age: '0' });
                         }
@@ -1196,6 +1197,26 @@ describe('Form:', sandbox(function () {
                                 .should.have.been.calledOnce;
 
                             instance.options.setValues
+                                .should.have.been.calledWith(
+                                    sinon.match.instanceOf(IncomingMessage),
+                                    sinon.match.instanceOf(ServerResponse),
+                                    sinon.match.func
+                                );
+
+                            next(err);
+                        });
+                });
+
+                it('calls preRedirect if the form is valid', function (next) {
+
+                    request(app).post('/form')
+                        .send({ firstName: 'Foo', lastName: 'Bar', age: '' })
+                        .end(function (err) {
+
+                            instance.options.preRedirect
+                                .should.have.been.calledOnce;
+
+                            instance.options.preRedirect
                                 .should.have.been.calledWith(
                                     sinon.match.instanceOf(IncomingMessage),
                                     sinon.match.instanceOf(ServerResponse),